requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.0
Pillow>=7.0.0
matplotlib>=3.1.0
tqdm>=4.50.0
